Django/jejuya/requirements.txt
apscheduler==3.11.0
colorama==0.4.6
Django==5.1.3
haversine==2.8.1
protobuf==5.28.3
pyproj==3.7.0
Requests==2.32.3
django-cors-headers==4.4.0
django-apscheduler==0.7.0
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0